dispatchers: add initial skeleton
Once the images have been downloaded it should be possible to dispatch
the images to one or several dispatchers. This way we should be able to
upload the images to an image service such as glance and, for example,
archive them for later inspection to a "file" dispatcher.

This commit provides the initial structure for this, together with a
"noop" dispatcher that will do nothing.