image source: pass extra arguments as **kwargs
We are passing the arguments for creating an image list source when
reading the yaml directly as kwargs. However, from the manager we do not
know what fields are required to create an image list source, therefore
we pass only the arguments we know and then we pass the rest in the
kwargs dict.

closes #14