add module and methods for extrating disks from OVA
OVA files can be used to distribute OVF containers. Add methods for
extrating the image disk, kernel and ramdisk files so that other modules
(like the glance dispatcher) can make use of it. Moreover, add a ovf.py
module that we can use to extract OVF descriptors and disk files.

Closes #17